Fixed warnings, use int for drawing functions

+++ src/capture_rectangle.cpp
@@ -70,8 +70,8 @@ CaptureRectangle::draw(SceneContext& sc)
         }
 
       sc.color().print_center(font, 
-                              pingu->get_center_pos().x,
-                              pingu->get_center_pos().y - 32,
+                              (int)pingu->get_center_pos().x,
+                              (int)pingu->get_center_pos().y - 32,
                               action_str,
                               1000);
     }
