+++ src/start_screen.cpp
@@ -71,7 +71,7 @@ public:
 
   void draw(DrawingContext& gc) {
     SurfaceButton::draw(gc);
-    gc.print_center(Fonts::chalk_normal, (float)x_pos + 30, (float)y_pos - 20, _("Ok"));
+    gc.print_center(Fonts::chalk_normal, x_pos + 30, y_pos - 20, _("Ok"));
   }
 
   bool is_at(int x, int y) {
@@ -112,7 +112,7 @@ public:
 
   void draw(DrawingContext& gc) {
     SurfaceButton::draw(gc);
-    gc.print_center(Fonts::chalk_normal, (float)x_pos + 55, (float)y_pos, _("Abort"));
+    gc.print_center(Fonts::chalk_normal, x_pos + 55, y_pos, _("Abort"));
   }
 
   void on_click() {
@@ -150,30 +150,30 @@ StartScreenComponent::draw(DrawingContext& gc)
 
   gc.print_center(Fonts::chalk_large,
                   gc.get_width()/2,
-                  (float)Display::get_height()/2 - 200,
+                  Display::get_height()/2 - 200,
                   _(plf.get_levelname()));
 
   gc.print_left(Fonts::chalk_normal,
-                (float)Display::get_width()/2 - 290,
-                (float)Display::get_height()/2 - 140,
+                Display::get_width()/2 - 290,
+                Display::get_height()/2 - 140,
                 format_description(800 - 230));
 
   y += 32;
 
-  gc.print_left (Fonts::chalk_normal, (float)left_x,  (float)y, _("Number of Pingus: "));
-  gc.print_right(Fonts::chalk_normal, (float)right_x, (float)y, StringUtil::to_string(plf.get_number_of_pingus()));
+  gc.print_left (Fonts::chalk_normal, left_x,  y, _("Number of Pingus: "));
+  gc.print_right(Fonts::chalk_normal, right_x, y, StringUtil::to_string(plf.get_number_of_pingus()));
 
-  gc.print_left (Fonts::chalk_normal, (float)left_x,  float(y += 30), _("Number to Save: "));
-  gc.print_right(Fonts::chalk_normal, (float)right_x, (float)y, StringUtil::to_string(plf.get_number_to_save()));
+  gc.print_left (Fonts::chalk_normal, left_x,  (y += 30), _("Number to Save: "));
+  gc.print_right(Fonts::chalk_normal, right_x, y, StringUtil::to_string(plf.get_number_to_save()));
 
-  gc.print_left (Fonts::chalk_normal, (float)left_x,  float(y += 30), _("Time: "));
-  gc.print_right(Fonts::chalk_normal, (float)right_x, (float)y, time_str);
+  gc.print_left (Fonts::chalk_normal, left_x,  (y += 30), _("Time: "));
+  gc.print_right(Fonts::chalk_normal, right_x, y, time_str);
 
-  gc.print_left (Fonts::chalk_normal, (float)left_x,  float(y += 30), _("Difficulty:"));
-  gc.print_right(Fonts::chalk_normal, (float)right_x, (float)y, StringUtil::to_string(plf.get_difficulty()) + "/100");
+  gc.print_left (Fonts::chalk_normal, left_x,  (y += 30), _("Difficulty:"));
+  gc.print_right(Fonts::chalk_normal, right_x, y, StringUtil::to_string(plf.get_difficulty()) + "/100");
 
-  gc.print_center(Fonts::chalk_small, (float)Display::get_width()/2,
-                  (float)Display::get_height()/2 + 270, _("Author: ") + plf.get_author());
+  gc.print_center(Fonts::chalk_small, Display::get_width()/2,
+                  Display::get_height()/2 + 270, _("Author: ") + plf.get_author());
 
   if (maintainer_mode)
     gc.print_left(Fonts::chalk_small, 110, 430, _("Filename: ") + plf.get_resname());
